Refactor transformers depth estimation class
Removed remnants from the tensorflow class, added a utils folder with methods and variables used by both classes, and tested the module\'s features
