+++ src/DepthEstimation/index.js
@@ -1,4 +1,4 @@
-// Copyright (c) 2019-2025 ml5
+// Copyright (c) 2019-2026 ml5
 //
 // This software is released under the MIT License.
 // https://opensource.org/licenses/MIT
@@ -11,9 +11,9 @@ import { handleModelName } from "../utils/handleOptions";
 /**
  * Creates a new DepthEstimation instance.
  * @param {string} [modelName="ARPortraitDepth"] - Model name.
- * @param {DepthEstimationOptions & DepthEstimationRuntimeOptions} [options] - Options.
+ * @param {TensorflowDepthEstimationOptions & TensorflowDepthEstimationRuntimeOptions | TransformersDepthEstimationRuntimeOptions} [options] - Options.
  * @param {function(): void} [callback] - Callback when loaded.
- * @returns {DepthEstimation} A new depthEstimation instance.
+ * @returns {TensorflowDepthEstimation | TransformersDepthEstimation} A new depthEstimation instance.
  */
 const depthEstimation = (...inputs) => {
   const { string, options = {}, callback } = handleArguments(...inputs);
